% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvecg_pkg.R
\docType{package}
\name{solvecg}
\alias{solvecg}
\title{Conjugate gradient method for solving and inverting}
\description{
The package contains two functions. Both of them are based on
the conjugate gradient method used for solving a symmetric system of equations.
solve_cg function solves iteratively a system of the form Ax = b.
invert_cg function calculates a matrix inverse.
}
